+++ nodeconductor/structure/views.py
@@ -654,16 +654,37 @@ class ProjectPermissionViewSet(viewsets.ModelViewSet):
     filter_backends = (filters.GenericRoleFilter, DjangoFilterBackend,)
     filter_class = filters.ProjectPermissionFilter
 
-    def can_manage_roles_for(self, project, role):
+    def can_manage_roles_for(self, project, role, expiration_time):
         user = self.request.user
         if user.is_staff:
             return True
 
-        if project.customer.has_user(user, models.CustomerRole.OWNER):
-            return True
+        customer_perm = models.CustomerPermission.objects.filter(
+            role=models.CustomerRole.OWNER,
+            user=user,
+            customer=project.customer,
+        )
+        if customer_perm.exists():
+            perm_expiration_time = customer_perm.first().expiration_time
+            if perm_expiration_time is None:
+                return True
+            elif expiration_time is not None and perm_expiration_time > expiration_time:
+                return True
 
-        if role == models.ProjectRole.ADMINISTRATOR and project.has_user(user, models.ProjectRole.MANAGER):
-            return True
+        if role != models.ProjectRole.ADMINISTRATOR:
+            return False
+
+        project_perm = models.ProjectPermission.objects.filter(
+            role=models.ProjectRole.MANAGER,
+            user=user,
+            project=project,
+        )
+        if project_perm.exists():
+            perm_expiration_time = project_perm.first().expiration_time
+            if perm_expiration_time is None:
+                return True
+            elif expiration_time is not None and perm_expiration_time > expiration_time:
+                return True
 
         return False
 
@@ -713,8 +734,9 @@ class ProjectPermissionViewSet(viewsets.ModelViewSet):
         affected_project = serializer.validated_data['project']
         affected_user = serializer.validated_data['user']
         role = serializer.validated_data['role']
+        expiration_time = serializer.validated_data.get('expiration_time')
 
-        if not self.can_manage_roles_for(affected_project, role):
+        if not self.can_manage_roles_for(affected_project, role, expiration_time):
             raise PermissionDenied('You do not have permission to perform this action.')
 
         if not affected_project.customer.get_users().filter(pk=affected_user.pk).exists():
@@ -725,8 +747,9 @@ class ProjectPermissionViewSet(viewsets.ModelViewSet):
     def perform_update(self, serializer):
         affected_project = serializer.instance.project
         role = serializer.instance.role
+        expiration_time = serializer.validated_data.get('expiration_time', serializer.instance.expiration_time)
 
-        if not self.can_manage_roles_for(affected_project, role)\
+        if not self.can_manage_roles_for(affected_project, role, expiration_time)\
                 or serializer.instance.user == self.request.user:
             raise PermissionDenied('You do not have permission to perform this action.')
 
@@ -736,8 +759,9 @@ class ProjectPermissionViewSet(viewsets.ModelViewSet):
         affected_user = instance.user
         affected_project = instance.project
         role = instance.role
+        expiration_time = instance.expiration_time
 
-        if not self.can_manage_roles_for(affected_project, role):
+        if not self.can_manage_roles_for(affected_project, role, expiration_time):
             raise PermissionDenied('You do not have permission to perform this action.')
 
         affected_project.remove_user(affected_user, role)
@@ -771,13 +795,22 @@ class CustomerPermissionViewSet(viewsets.ModelViewSet):
     )
     filter_class = filters.CustomerPermissionFilter
 
-    def can_manage_roles_for(self, customer):
+    def can_manage_roles_for(self, customer, expiration_time):
         user = self.request.user
         if user.is_staff:
             return True
 
-        if customer.has_user(user, models.CustomerRole.OWNER):
-            return True
+        customer_perm = models.CustomerPermission.objects.filter(
+            role=models.CustomerRole.OWNER,
+            user=user,
+            customer=customer,
+        )
+        if customer_perm.exists():
+            perm_expiration_time = customer_perm.first().expiration_time
+            if perm_expiration_time is None:
+                return True
+            elif expiration_time is not None and perm_expiration_time > expiration_time:
+                return True
 
         return False
 
@@ -838,8 +871,9 @@ class CustomerPermissionViewSet(viewsets.ModelViewSet):
     def perform_create(self, serializer):
         affected_customer = serializer.validated_data['customer']
         affected_user = serializer.validated_data['user']
+        expiration_time = serializer.validated_data.get('expiration_time')
 
-        if not self.can_manage_roles_for(affected_customer):
+        if not self.can_manage_roles_for(affected_customer, expiration_time):
             raise PermissionDenied('You do not have permission to perform this action.')
 
         if not affected_customer.get_users().filter(pk=affected_user.pk).exists():
@@ -852,8 +886,9 @@ class CustomerPermissionViewSet(viewsets.ModelViewSet):
 
     def perform_update(self, serializer):
         affected_customer = serializer.instance.customer
+        expiration_time = serializer.validated_data.get('expiration_time', serializer.instance.expiration_time)
 
-        if not self.can_manage_roles_for(affected_customer) \
+        if not self.can_manage_roles_for(affected_customer, expiration_time) \
                 or serializer.instance.user == self.request.user:
             raise PermissionDenied('You do not have permission to perform this action.')
 
@@ -863,8 +898,9 @@ class CustomerPermissionViewSet(viewsets.ModelViewSet):
         affected_user = instance.user
         affected_customer = instance.customer
         role = instance.role
+        expiration_time = instance.expiration_time
 
-        if not self.can_manage_roles_for(affected_customer):
+        if not self.can_manage_roles_for(affected_customer, expiration_time):
             raise PermissionDenied('You do not have permission to perform this action.')
 
         affected_customer.remove_user(affected_user, role)